requirements.txt
discord.py
requests
python-decouple
datetime
asyncio
translate
unidecode
pytz
apscheduler
repackage